requirements.txt
Flask==3.1.2
groq==1.0.0
python-dotenv==1.2.1
requests==2.32.5
